Changed the layout to display a gradient header with images related to the topic Games.
Added a a set of initial buttons for the user to test.
Added a feature that adds a different button color when new topic is included.
changed the size of the gifs displayed
centered all the buttons being added.
centered and resized the search bar

Pending:
resize of buttons when in phone device view.
resize of gifs to full width when in phone device view.
resize and alignment of search bar when in phone device view.